// src/lib/shopify/types.ts
export type Menu = {
  title: string;
  path: string;
  items: {
    title: string;
    path: string;
  }[];
};